+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.1.0 2016-09-04
+
+### Added
+- pattern to JSONString 
+- additionalProperties to JSONObject
+
+
 ## 1.0.0 2016-09-03
 
 ### Added